Bug fix.
A bug was found in the code which prevented the app from printing Hello Joe Bloggs. With the use of the call backs the program creates an asynchronous non-blocking call where the res.send(response) is sent to the client before the “response = response + name” join of the strings. This is fixed by sending the updated response variable to the client, in the scope where the concatenation of the response variable is carried out. Note that this is a temporary fix to the issue.